third_party/WebKit/LayoutTests/fast/canvas/canvas-toBlob-file-vs-blob.html

<script src = "../../resources/js-test.js"></script>
<script type = "text/javascript"> 
if (window.testRunner)
{
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

description("Test if toBlob creates a Blob-type object.");

var resultImageObject = null;
canvas = document.createElement('canvas');
canvas.toBlob(function(b) {
    resultImageObject = b;
    shouldBeFalse('resultImageObject instanceof File');
    shouldBeTrue('resultImageObject instanceof Blob');
    if (window.testRunner)
        testRunner.notifyDone();
});
</Script>
