third_party/WebKit/LayoutTests/fast/canvas/canvas-toBlob-file-vs-blob-expected.txt

Test if toBlob creates a Blob-type object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS resultImageObject instanceof File is false
PASS resultImageObject instanceof Blob is true
